+++ include/Sequence.hpp
@@ -8,13 +8,13 @@ class Sequence
 
         virtual T GetFirst() = 0;
         virtual T GetLast() = 0;
-        virtual T Get(int index) = 0;
-        virtual Sequence<T>* GetSubsequence(int startIndex, int endIndex) = 0;
-        virtual int GetLength() = 0;
+        virtual T Get(size_t index) = 0;
+        virtual Sequence<T>* GetSubsequence(size_t startIndex, size_t endIndex) = 0;
+        virtual size_t GetLength() = 0;
 
         virtual Sequence<T>* Append(T item) = 0;
         virtual Sequence<T>* Prepend(T item) = 0;
-        virtual Sequence<T>* InsertAt(T item, int index) = 0;
+        virtual Sequence<T>* InsertAt(T item, size_t index) = 0;
         virtual Sequence<T>* Concat(Sequence<T>* list) = 0;
 
         virtual Sequence<T>* Map(T (*)(T)) = 0;
